// src/api-client/models/promotion-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spring-Adam-Store
 * Api Service
 *
 * The version of the OpenAPI document: v1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface PromotionResponse
 */
export interface PromotionResponse {
    /**
     * 
     * @type {number}
     * @memberof PromotionResponse
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof PromotionResponse
     */
    'code'?: string;
    /**
     * 
     * @type {string}
     * @memberof PromotionResponse
     */
    'description'?: string;
    /**
     * 
     * @type {number}
     * @memberof PromotionResponse
     */
    'discountPercent'?: number;
    /**
     * 
     * @type {string}
     * @memberof PromotionResponse
     */
    'startDate'?: string;
    /**
     * 
     * @type {string}
     * @memberof PromotionResponse
     */
    'endDate'?: string;
    /**
     * 
     * @type {string}
     * @memberof PromotionResponse
     */
    'status'?: PromotionResponseStatusEnum;
    /**
     * 
     * @type {string}
     * @memberof PromotionResponse
     */
    'createdBy'?: string;
    /**
     * 
     * @type {string}
     * @memberof PromotionResponse
     */
    'createdAt'?: string;
}

export const PromotionResponseStatusEnum = {
    Active: 'ACTIVE',
    Inactive: 'INACTIVE'
} as const;

export type PromotionResponseStatusEnum = typeof PromotionResponseStatusEnum[keyof typeof PromotionResponseStatusEnum];