// src/api-client/models/api-response-page-response-promotion-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spring-Adam-Store
 * Api Service
 *
 * The version of the OpenAPI document: v1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { PageResponsePromotionResponse } from './page-response-promotion-response';

/**
 * 
 * @export
 * @interface ApiResponsePageResponsePromotionResponse
 */
export interface ApiResponsePageResponsePromotionResponse {
    /**
     * 
     * @type {number}
     * @memberof ApiResponsePageResponsePromotionResponse
     */
    'code'?: number;
    /**
     * 
     * @type {string}
     * @memberof ApiResponsePageResponsePromotionResponse
     */
    'message'?: string;
    /**
     * 
     * @type {PageResponsePromotionResponse}
     * @memberof ApiResponsePageResponsePromotionResponse
     */
    'result'?: PageResponsePromotionResponse;
}